Added new file for Coin Enumeration. Also modified price for items in VendingItem to int, to represent coins in x100 of actual values. Removed Drinks class as it is no longer needed. Implemented the return by id, delete by id and update by id for the VendingItems class under endpoint /items.

+++ src/main/java/com/rob/domain/VendingItem.java
@@ -14,6 +14,6 @@ public class VendingItem implements Serializable {
     private String id;
     private String title;
     private String description;
-    private Float price;
+    private int price; // Price in x100 increment i.e 100 is $1.00
 
 }
